Move ExpressionTokenizer to its own class
